Staretd testing auth route
Conducting unit testing on auth route.

+++ Routes/auth_routes.py
@@ -2,43 +2,59 @@ from flask import request, jsonify, make_response
 from Routes import Routes
 from Tools import GeneratorTools as gt, PasswordTools as pt, StringTools as st
 from Database import UserFactory, VerificationCodeFactory, SessionFactory, PasswordResetFactory
+from flask_limiter import Limiter
+from flask_limiter.util import get_remote_address
+from Decorators import is_user_authorized
+
+limiter = Limiter(
+    Routes,
+    key_func=get_remote_address,
+    default_limits=["200 per day", "50 per hour"]
+)
+
+# Don't limit requests coming from localhost
+
+@limiter.request_filter
+def ip_whitelist():
+    return request.remote_addr == "127.0.0.1"
 
 CLIENT_DOMAIN = '127.0.0.1:5500'
 
 @Routes.route('/signup', methods=['POST'])
+@limiter.limit("6/hour")
 def register():
     
     if not request.headers.get('Content-Type'):
-        return jsonify({'error': True, 'message': 'Missing required headers!', 'code': 'MISSING_REQUIRED_HEADERS'})
+        return jsonify({'error': True, 'message': 'Missing required headers!', 'code': 'MISSING_REQUIRED_HEADERS'}), 400
     
     jsonData = request.get_json()
 
     if not jsonData or 'email' not in jsonData or 'name' not in jsonData or 'password' not in jsonData:
-        return jsonify({'error': True, 'message': 'Not accepted!', 'code': 'MISSING_REQUIRED_KEYS'})
+        return jsonify({'error': True, 'message': 'Not accepted!', 'code': 'MISSING_REQUIRED_KEYS'}), 400
 
     if not jsonData['email'] or not jsonData['name'] or not jsonData['password']:
-        return jsonify({'error': True, 'message': 'Empty values in input field', 'code': 'MISSING_REQUIRED_VALUES'})
+        return jsonify({'error': True, 'message': 'Empty values in input field', 'code': 'MISSING_REQUIRED_VALUES'}), 400
     
     if not isinstance(jsonData['email'], str) or not isinstance(jsonData['name'], str) or not isinstance(jsonData['password'], str):
-        return jsonify({'error': True, 'message': 'Recieved invalid value type!', 'code': 'INVALID_TYPE'})
+        return jsonify({'error': True, 'message': 'Recieved invalid value type!', 'code': 'INVALID_TYPE'}), 400
     
     emailCheck = st(jsonData['email'])
     if not emailCheck.is_email_valid():
-        return jsonify({'error': True, 'message': 'Email is not valid!', 'code': 'INVALID_EMAIL'})
+        return jsonify({'error': True, 'message': 'Email is not valid!', 'code': 'INVALID_EMAIL'}), 400
 
     nameCheck = st(jsonData['name'])
     if not nameCheck.is_name_valid():
-        return jsonify({'error': True, 'message': 'Name is not valid!', 'code': 'INVALID_NAME'})
+        return jsonify({'error': True, 'message': 'Name is not valid!', 'code': 'INVALID_NAME'}), 400
 
     passwordString = pt(jsonData['password'])
     if not passwordString.is_password_strong():
-        return jsonify({'error': True, 'message': 'Password needs to be stronger!', 'code': 'INVALID_PASSWORD'})
+        return jsonify({'error': True, 'message': 'Password needs to be stronger!', 'code': 'WEAK_PASSWORD'}), 400
 
     try:
         user = UserFactory()
         verification = VerificationCodeFactory()
     except:
-        return jsonify({'error': True, 'message': 'Unable to initialize database!', 'code': 'DB_TABLE_ERROR'})
+        return jsonify({'error': True, 'message': 'Unable to initialize!', 'code': 'INIT_ERROR'})
 
     if user.userAlreadyExistsInDB(jsonData['email']):
         return jsonify({'error': True, 'message': 'User already exists!', 'code': 'DUPLICATE_USER'})
@@ -55,12 +71,14 @@ def register():
         verification.saveVerificationCode(jsonData['email'], verificationCode)
     except:
         user.deleteUser(jsonData['email'])
+        verification.deleteVerificationCode(jsonData['email'])
         return jsonify({'error': True, 'message': 'Unable to initialize database!', 'code': 'DB_TABLE_ERROR'})
     
     return jsonify({'error': False, 'message': 'User created successfully!', 'code': 'SUCCESS'})
 
 
 @Routes.route('/verify', methods=['POST'])
+@limiter.limit("20/hour")
 def verify():
 
     if not request.headers.get('Content-Type'):
@@ -89,7 +107,7 @@ def verify():
     if not verificationCodeInDb:
         return jsonify({'error': True, 'message': 'Verification code for this user does not exist!', 'code': 'DOES_NOT_EXIST'})
     
-    # Check if provide verification code matches with the one in the database.
+    # Check if provide verification code does not match with the one in the database.
     if verificationCodeInDb['verificationCode'] != jsonData['verificationCode']:
         return jsonify({'error': True, 'message': 'Code does not match!', 'code': 'DOES_NOT_MATCH'})
     
@@ -113,6 +131,7 @@ def verify():
     
 
 @Routes.route('/login', methods=['POST'])
+@limiter.limit("7 / 2 minute")
 def login():
     
     if not request.headers.get('Content-Type'):
@@ -241,8 +260,9 @@ def logout():
     return res
     
 
-@Routes.route('/reset-password-initial', methods=['POST'])
-def reset_password_initial():
+@Routes.route('/reset-password-generate-token', methods=['POST'])
+@limiter.limit("7 / 2 minute")
+def reset_password_generate_token():
     if not request.headers.get('Content-Type'):
         return jsonify({'error': True, 'message': 'Forbidden. Invalid content-type'}), 403
     
@@ -264,16 +284,6 @@ def reset_password_initial():
         return jsonify({'error': True, 'message': 'User does not exist!'})
     
     randomToken = gt.password_reset_token_generator()
-    tokenData = passwordResetFactory.getPasswordResetTokenData()
-    if tokenData['token']:
-        if passwordResetFactory.isTokenExpired():
-            try:
-                passwordResetFactory.updateTokenExpiration(randomToken)
-                return jsonify({'error': False, 'message': 'Token update.'})
-            except:
-                return jsonify({'error': True, 'message': 'Unable to update token.'})
-        else:
-            return jsonify({'error': True, 'message': 'Token is not expired yet. Please check your email'})
     
     try:
         passwordResetFactory.savePasswordResetToken(randomToken)
@@ -305,18 +315,18 @@ def reset_password(token):
     except:
         return jsonify({'error': True, 'message': 'Unable to initialize'})
     
+    if not user.userAlreadyExistsInDB(jsonData['email']):
+        return jsonify({'error': True, 'message': 'Hmmm... User not found!'})
+    
     tokenData = passwordResetFactory.getPasswordResetTokenData()
     if not tokenData['token']:
         return jsonify({'error': True, 'message': 'User has not initiated a password reset!'})
 
     if passwordResetFactory.isTokenExpired():
-        return jsonify({'error': True, 'message': 'Token has expired! Please resubmit anouther request!'})
+        return jsonify({'error': True, 'message': 'Token has expired! Please resubmit another request to reset password!'})
 
     if tokenData['token'] != token:
         return jsonify({'error': True, 'message': 'Token doesn\'t match!'})
-
-    if not user.userAlreadyExistsInDB(jsonData['email']):
-        return jsonify({'error': True, 'message': 'Hmmm... User not found!'})
 
     hashedPassword = passwordHandler.encrypt_password()
     try:
@@ -328,37 +338,66 @@ def reset_password(token):
     return jsonify({'error': False, 'message': 'Password changed successfully'})
     
 
-@Routes.route('/delete-user', methods=['DELETE'])
-def delete_user():
+#TODO
+@Routes.route('/request-new-verification-code', methods=['POST'])
+def request_new_verification_code():
+    pass
 
+
+#TODO
+@Routes.route('/change-user-password', methods=['POST'])
+@is_user_authorized
+def change_user_password():
+    pass
+
+
+@Routes.route('/delete-user', methods=['DELETE'])
+@is_user_authorized
+def delete_user():
+    
     jsonData = request.get_json()
-    if not 'email' in jsonData or not jsonData['email'] or not jsonData:
-        return jsonify({'error': True, 'message': 'Missing data. Request Not Acceptable!'})
-    if not 'password' in jsonData or not jsonData['password'] or not jsonData:
-        return jsonify({'error': True, 'message': 'Missing data. Request Not Acceptable!'})
+
+    if not 'email' in jsonData or not jsonData['email']:
+        return jsonify({'error': True, 'message': 'Missing data. Not Acceptable'})
+    if not 'password' in jsonData or not jsonData['password']:
+        return jsonify({'error': True, 'message': 'Missing data. Not Acceptable'})
+
+    if not request.cookies.get('appSessionId'):
+            return jsonify({'error': True, 'message': 'Cookies are missing. Not authorized!'})
+   
+    appSessionId = request.cookies.get('appSessionId')
     
     try:
+        session = SessionFactory()
         user = UserFactory()
+        verification = VerificationCodeFactory()
     except:
         return jsonify({'error': True, 'message': 'Cannot initialize user table!'})
+    
+    if not session.sessionIdExists(appSessionId) or session.sessionHasExpired(appSessionId):
+        return jsonify({'error': True, 'message': 'Invalid session. Please log in again!', 'code': "INVALID_SESSION"})
+    
+    retrievedUserEmail = session.getSessionData(appSessionId)['userEmail']
+    
+    if retrievedUserEmail != jsonData['email']:
+        return jsonify({'error': True, 'message': "Given user email and user session email does not match!"})
     
     if not user.userAlreadyExistsInDB(jsonData['email']):
         return jsonify({'error': True, 'message': 'User does not exist!'})
 
     passwordFromDB = user.getUserPassword(jsonData['email'])[0][0]
     passwordHandler = pt(jsonData['password'].encode("utf-8"))
-    try:
-        if not passwordHandler.check_password(passwordFromDB.encode("utf-8")):
-            return jsonify({'error': True, 'message': 'Iccorrect password. Could not delete user!'})
-    except:
-        return jsonify({'error': True, 'message': 'Unable to delete user. Try resetting password and then delete your account!'})
+    if not passwordHandler.check_password(passwordFromDB.encode("utf-8")):
+        return jsonify({'error': True, 'message': 'Iccorrect password. Could not delete user!'})
     
     try:
         user.deleteUser(jsonData['email'])
+        session.deleteSession(appSessionId)
+        verification.deleteVerificationCode(jsonData['email'])
+        
     except:
         return jsonify({'error': True, 'message': 'Can\'t delete user!'})
 
-    
     listOfCookies = [
         {
             'key': "appSessionId", 
@@ -390,10 +429,25 @@ def delete_user():
         )
     return res
 
-#TODO
-@Routes.route('/request-new-verification-code', methods=['POST'])
-def request_new_verification_code():
-    pass
+
+@Routes.errorhandler(404)
+def route_not_found(e):
+    return jsonify({"error": True, "message": "The requested url was not found!", "code": "NOT_FOUND"})
+
+
+@Routes.errorhandler(429)
+def ratelimit_handler(e):
+    return jsonify({"error": True, "message": "Too many requests. Please try again later.", "code": "TOO_MANY_REQUESTS"}), 429
+
+
+@Routes.errorhandler(500)
+def server_error_handler(e):
+    return jsonify({"error": True, "message": "Internal server error!", "code": "SERVER_ERROR"}), 500
+
+
+@Routes.errorhandler(405)
+def method_not_allowed_handler(e):
+    return jsonify({"error": True, "message": "Method rejected by server!", "code": "REJECTED"}), 405
 
 
 def getPublicIpAddressOfClient():
